code-quality: move report genration utils to common package
