+++ packages/cli-dashboard/src/components/utils/reportDownloader/generateSiteMapReportandDownload.ts
@@ -19,11 +19,11 @@
  */
 import JSZip from 'jszip';
 import { saveAs } from 'file-saver';
+import type { CompleteJson } from '@ps-analysis-tool/common';
 
 /**
  * Internal dependencies
  */
-import type { CompleteJson } from '../../../types';
 import { createZip, getFolderName } from './utils';
 
 const generateSiteMapReportandDownload = async (JSONReport: CompleteJson[]) => {
