+++ packages/cli-dashboard/src/types.ts
@@ -16,7 +16,6 @@
 /**
  * External dependencies
  */
-import type { TechnologyData } from '@ps-analysis-tool/common';
 import { type Cookie as ParsedCookie } from 'simple-cookie';
 
 export type SanitisedCookieType = ParsedCookie & {
@@ -33,44 +32,3 @@ export interface SingleTechnology {
   website: string;
   categories: string;
 }
-
-export type CookieJsonDataType = {
-  name: string;
-  value: string;
-  domain: string;
-  partitionKey?: string;
-  path: string;
-  expires: string;
-  httpOnly: boolean;
-  secure: boolean;
-  sameSite: string;
-  platform: string;
-  category: string;
-  description: string;
-  isFirstParty: boolean;
-  pageUrl: string;
-  requestUrls?: { [id: string]: string };
-  frameUrls?: { [id: string]: string };
-  isBlocked: boolean;
-  blockedReasons?: string[];
-  GDPR?: string;
-};
-
-export type CookieFrameStorageType = {
-  [frame: string]: {
-    [cookieKey: string]: CookieJsonDataType;
-  };
-};
-
-export type CompleteJson = {
-  pageUrl: string;
-  cookieData: {
-    [frame: string]: {
-      cookiesCount: number;
-      frameCookies: {
-        [cookieKey: string]: CookieJsonDataType;
-      };
-    };
-  };
-  technologyData: TechnologyData[];
-};
