+++ packages/cli-dashboard/tsconfig.json
@@ -6,7 +6,8 @@
   "include": [
     "src",
     "../../node_modules/@types/jest/index.d.ts",
-    "src/**/*.json"
+    "src/**/*.json",
+    "../common/src/utils/generateReports"
   ],
   "references": [
     { "path": "../common" },
