+++ packages/common/src/index.ts
@@ -32,5 +32,6 @@ export { default as parseResponseReceivedExtraInfo } from './utils/parseResponse
 export { default as parseRequestWillBeSentExtraInfo } from './utils/parseRequestWillBeSentExtraInfo';
 export { default as getDomainFromUrl } from './utils/getDomainFromUrl';
 export { default as noop } from './utils/noop';
+export * from './utils/generateReports';
 export * from './cookies.types';
 export const UNKNOWN_FRAME_KEY = 'Unknown Frame(s)';
