+++ packages/common/src/utils/generateReports/generateSummaryDataCSV.ts
@@ -17,7 +17,7 @@
 /**
  * Internal dependencies
  */
-import type { CompleteJson, CookieJsonDataType } from '../../../../types';
+import type { CompleteJson, CookieJsonDataType } from '../../cookies.types';
 
 const generateSummaryDataCSV = (siteAnalysisData: CompleteJson): string => {
   const frameCookieDataMap = siteAnalysisData.cookieData;
