+++ packages/common/src/utils/generateReports/tests/generateAffectedCookiesCSV.ts
